requirements.txt
django
django-cors-headers
djangorestframework
drf-spectacular
flask
flasgger
flask-marshmallow
Flask-SQLAlchemy
flask-swagger-ui
marshmallow-sqlalchemy
pillow
pyotp
python-dotenv
PyJWT
qrcode
stripe
